Add files via upload
Done a project which includes processes like Registration and Login of the student's details, compiler page, Leave Letter and materials where you can download and admin page which assigning the task based on their course., mark and status of the task. It stores all the data in the database which makes it easy to fetch and update whenever needed.